Rename EntityResourceCollection files.
