+++ packages/core/src/common/services/entity-resource-collection.service.ts
