+++ packages/core/src/common/services/index.ts
@@ -1,2 +1,2 @@
 export { IResourceCollection } from './resource-collection.interface';
-export { EntityResourceCollection } from './entity-serializer.service';
+export { EntityResourceCollection } from './entity-resource-collection.service';
